+++ core/execution/future/liquidity_provision_test.go
@@ -87,12 +87,10 @@ func TestSubmit(t *testing.T) {
 		mktCfg := getMarket(pMonitorSettings, &types.AuctionDuration{
 			Duration: 10000,
 		})
-		mktCfg.Fees = &types.Fees{
-			Factors: &types.FeeFactors{
-				LiquidityFee:      num.DecimalFromFloat(0.001),
-				InfrastructureFee: num.DecimalFromFloat(0.0005),
-				MakerFee:          num.DecimalFromFloat(0.00025),
-			},
+		mktCfg.Fees.Factors = &types.FeeFactors{
+			LiquidityFee:      num.DecimalFromFloat(0.001),
+			InfrastructureFee: num.DecimalFromFloat(0.0005),
+			MakerFee:          num.DecimalFromFloat(0.00025),
 		}
 		mktCfg.TradableInstrument.RiskModel = &types.TradableInstrumentLogNormalRiskModel{
 			LogNormalRiskModel: &types.LogNormalRiskModel{
@@ -248,12 +246,11 @@ func TestSubmit(t *testing.T) {
 		mktCfg := getMarket(pMonitorSettings, &types.AuctionDuration{
 			Duration: 10000,
 		})
-		mktCfg.Fees = &types.Fees{
-			Factors: &types.FeeFactors{
-				InfrastructureFee: num.DecimalFromFloat(0.0005),
-				MakerFee:          num.DecimalFromFloat(0.00025),
-			},
+		mktCfg.Fees.Factors = &types.FeeFactors{
+			InfrastructureFee: num.DecimalFromFloat(0.0005),
+			MakerFee:          num.DecimalFromFloat(0.00025),
 		}
+
 		mktCfg.TradableInstrument.RiskModel = &types.TradableInstrumentLogNormalRiskModel{
 			LogNormalRiskModel: &types.LogNormalRiskModel{
 				RiskAversionParameter: num.DecimalFromFloat(0.001),
@@ -335,11 +332,9 @@ func TestAmend(t *testing.T) {
 		mktCfg := getMarket(defaultPriceMonitorSettings, &types.AuctionDuration{
 			Duration: 10000,
 		})
-		mktCfg.Fees = &types.Fees{
-			Factors: &types.FeeFactors{
-				InfrastructureFee: num.DecimalFromFloat(0.0005),
-				MakerFee:          num.DecimalFromFloat(0.00025),
-			},
+		mktCfg.Fees.Factors = &types.FeeFactors{
+			InfrastructureFee: num.DecimalFromFloat(0.0005),
+			MakerFee:          num.DecimalFromFloat(0.00025),
 		}
 		mktCfg.TradableInstrument.RiskModel = &types.TradableInstrumentLogNormalRiskModel{
 			LogNormalRiskModel: &types.LogNormalRiskModel{